backend/kalshi_executor.py
"""
kalshi_executor.py
This module handles the execution of trade signals for the Kalshi prediction market.
For the demo, it simulates trade execution by printing signals to the console.
"""

import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def execute_trade_signal(trade_signal: dict):
    """
    Simulates trade execution based on the provided trade signal.
    In a production environment, this would make API calls to Kalshi.
    
    Args:
        trade_signal (dict): A dictionary containing trade signal information
            with keys: 'ticker', 'side', 'conviction', 'explanation', etc.
    
    Returns:
        None
    """
    try:
        print(f"Received trade signal: {trade_signal}")
        
        # Only execute trade if conviction is high enough
        if trade_signal['conviction'] > 0.9:
            print("-----------------------------------------")
            print(f"SIMULATED TRADE PLACED:")
            print(f"  Ticker: {trade_signal['ticker']}")
            print(f"  Side: {trade_signal['side']}")
            print(f"  Forecast Temperature: {trade_signal['forecast_temp']}°F")
            print(f"  Conviction Score: {trade_signal['conviction']:.2f}")
            print(f"  Timestamp: {trade_signal['timestamp']}")
            print(f"  Reason: {trade_signal['explanation']}")
            print("-----------------------------------------")
        else:
            print(f"Conviction score {trade_signal['conviction']:.2f} too low, no trade placed.")
            
    except Exception as e:
        print(f"Failed to execute trade signal: {e}")
        
    return None
