backend/requirements.txt
# Dependencies for Kalshi Weather Bot
pathway
streamlit
requests
langchain
langchain-openai
python-dotenv
